+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Incompatible type errors in the `PaginationManager::getCurrentPageIndex()` and `PaginationManager::getPageCount()` methods
 - Database connections are not properly released by the close function
+- Avoid returning sensitive information in the User instance
 
 ## [2.0.1] - 2024-11-23