+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Incompatible type errors in the `PaginationManager::getCurrentPageIndex()` and `PaginationManager::getPageCount()` methods
 - Database connections are not properly released by the close function
 - Avoid returning sensitive information in the User instance
+- Fixed a legacy bug in the DataValidator Twig extension
 
 ## [2.0.1] - 2024-11-23