+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- New `IsJson` data validator filter
+
+### Changed
+
+- Upgrade jQuery to version 4.0.0
+
 ## [2.1.0] - 2025-06-20
 
 ### Added